+++ qoa4ml_lib/exrequirements.txt
@@ -4,4 +4,4 @@ Pillow==10.0.0
 prometheus-client==0.16.0
 setuptools==65.5.1
 tensorflow==2.13.0
-ydata-quality==0.1.0
+ydata-quality
