+++ qoa4ml_lib/qoa4ml/probes/process_monitoring_probe.py
@@ -3,14 +3,14 @@ import time
 import os
 import psutil
 import importlib
-from qoa4ml.qoaUtils import (
+from ..qoaUtils import (
     convert_to_mbyte,
     report_proc_child_cpu,
     report_proc_mem,
     get_process_allowed_cpus,
     get_process_allowed_memory,
 )
-from .core.probe import Probe
+from .probe import Probe
 
 logging.basicConfig(
     format="%(asctime)s:%(levelname)s -- %(message)s", level=logging.INFO
