+++ qoa4ml_lib/qoa4ml/probes/system_monitoring_probe.py
@@ -1,18 +1,18 @@
 import os
 import time
 import importlib
-from qoa4ml.qoaUtils import (
+from ..qoaUtils import (
     get_sys_cpu_util,
     get_sys_mem,
     convert_to_gbyte,
     convert_to_mbyte,
     get_sys_cpu_metadata,
 )
-from qoa4ml.gpuUtils import (
+from ..gpuUtils import (
     get_sys_gpu_metadata,
     get_sys_gpu_usage,
 )
-from .core.probe import Probe
+from .probe import Probe
 
 NODE_NAME = os.getenv("NODE_NAME")
 if not NODE_NAME:
